chore: move parser and pgxv5 to public package
